[W3D2] CreateVoucher class 내용을 VoucherRepository 인터페이스에 포함시킴, percentDiscount를 다시 float->long 으로 변경

+++ src/main/java/org/prgrms/kdt/voucher/PercentDiscountVoucher.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 
 public class PercentDiscountVoucher implements Voucher {
     private final UUID voucherId;
-    private final float percent;
+    private final long percent;
 
-    public PercentDiscountVoucher(final UUID voucherId, final float percent) {
+    public PercentDiscountVoucher(final UUID voucherId, final long percent) {
         this.voucherId = voucherId;
         this.percent = percent;
     }
